+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file. We try to a
 ## 2.0.0-BETA-2-SNAPSHOT
 
 ### Added
+- Add "-E" command line option to expose all environment variable as top-level entry in the data model 
 - Support for [Grok](https://github.com/thekrakken/java-grok) as better alternative for regular expressions
 - Support for streaming textual content over `System.in` to allow processing of arbitrary large source files
 - Display Git version information when using `-V` command line option